ReadUSB: Fix compiler warning
ccid_usb.c: In function ‘ReadUSB’:
ccid_usb.c:967:61: warning: comparison of integer expressions of different signedness: ‘int’ and ‘unsigned int’ [-Wsign-compare]
  967 |                                 if (concurrent[slot].length > *length)

+++ src/ccid_usb.c
@@ -964,7 +964,7 @@ read_again:
 				slot);
 			if (concurrent[slot].length > 0)
 			{
-				if (concurrent[slot].length > *length)
+				if (concurrent[slot].length > (int)*length)
 					DEBUG_CRITICAL3("Received %d bytes but expected only %d",
 						concurrent[slot].length, *length);
 				else
